Update bufferLength once instead of twice per channel
Hello!

I came upon this project and adapted it for something else I'm working on, and ran into a bug where the encoded WAV file would be twice the length I expected.  I changed this in my adaptation to address, and figured there might be a similar issue upstream here.

Thanks for your work here and for sharing!